add new file common/ticket.go
